06 server

+++ client-admin/src/store/actions/genresAction.js
@@ -1,4 +1,5 @@
 import { SET_GENRES, BASE_URL } from "../actionTypes/actionTypes";
+import Swal from "sweetalert2";
 
 export function setGenres(data) {
   return {
@@ -10,7 +11,11 @@ export function setGenres(data) {
 export function fetchGenres() {
   return async (dispatch) => {
     try {
-      const response = await fetch("http://localhost:4000/genres");
+      const response = await fetch(BASE_URL + `/movies/genres`, {
+        headers: {
+          access_token: localStorage.getItem("access_token"),
+        },
+      });
 
       if (!response.ok) {
         throw new Error(`Something's wrong!`);
@@ -28,41 +33,75 @@ export function fetchGenres() {
 export function createGenre(genre) {
   return async function (dispatch) {
     try {
-      const response = await fetch(BASE_URL + `/genres`, {
+      let response = await fetch(BASE_URL + `/movies/genres`, {
         method: "POST",
         body: JSON.stringify(genre),
         headers: {
+          access_token: localStorage.getItem("access_token"),
           "Content-Type": "application/json",
         },
       });
 
       if (!response.ok) {
-        throw new Error(`Genre is failed to add`);
+        response = await response.json();
+        throw response.message;
       }
-      // const data = await response.json();
-      // console.log(data);
-      dispatch(fetchGenres())
+      successSwal("New genre is created successfully");
+      dispatch(fetchGenres());
     } catch (error) {
-      console.log(error);
+      errorSwal(error);
     }
   };
 }
 
-export const deleteGenre = (id) => {
-  console.log(id);
-  return async (dispatch) => {
+export function deleteGenre(id) {
+  return (dispatch) => {
     try {
-      await fetch(BASE_URL + "/categories/" + id, {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-          // access_token: localStorage.getItem(`access_token`),
-        },
-        // body: JSON.stringify(id)
+      confirmSwal().then(async (result) => {
+        if (result.isConfirmed) {
+          let response = await fetch(BASE_URL + `/movies/genres/${id}`, {
+            method: "DELETE",
+            headers: { access_token: localStorage.getItem("access_token") },
+          });
+          if (!response.ok) {
+            throw new Error("Internal Server Error");
+          }
+          Swal.fire("Deleted!", "Genre is deleted successfully", "success");
+          dispatch(fetchGenres());
+        }
       });
-      dispatch(fetchGenres());
-    } catch (error) {
-      console.log(error);
+    } catch (err) {
+      errorSwal(err);
     }
   };
-};
+}
+
+function successSwal(message) {
+  Swal.fire({
+    position: "center",
+    icon: "success",
+    title: message,
+    showConfirmButton: false,
+    timer: 1500,
+  });
+}
+
+function errorSwal(message) {
+  Swal.fire({
+    icon: "error",
+    title: "Oops...",
+    text: "Something went wrong!",
+  });
+}
+
+function confirmSwal() {
+  return Swal.fire({
+    title: "Are you sure?",
+    text: "You won't be able to revert this!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Yes, I'm sure!",
+  });
+}
